refactor: :recycle: BeerListFilterAndSorter, BeerListSearchResult prop 수정
